+++ components/nimbus/src/tests/stateful/test_gecko_prefs.rs
@@ -3,12 +3,12 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 use crate::{
-    enrollment::ExperimentEnrollment,
+    enrollment::{ExperimentEnrollment, PreviousGeckoPrefState},
     error::Result,
     json::PrefValue,
     stateful::gecko_prefs::{
         create_feature_prop_pref_map, GeckoPrefHandler, GeckoPrefState, GeckoPrefStore,
-        GeckoPrefStoreState,
+        GeckoPrefStoreState, PrefBranch, PrefEnrollmentData,
     },
     tests::helpers::{get_multi_feature_experiment, TestGeckoPrefHandler},
     EnrolledExperiment,
@@ -216,6 +216,149 @@ fn test_gecko_pref_store_pref_is_user_set() -> Result<()> {
 
     assert!(store.pref_is_user_set("test.pref.set_by_user"));
     assert!(!store.pref_is_user_set("test.pref.NOT.set_by_user"));
+
+    Ok(())
+}
+
+#[test]
+fn test_build_prev_gecko_pref_states() -> Result<()> {
+    let pref_state_1 = GeckoPrefState::new("test.some.pref.1", Some(PrefBranch::Default))
+        .with_gecko_value(json!("gecko-pref-value-1"))
+        .with_enrollment_value(PrefEnrollmentData {
+            experiment_slug: "experiment-slug-1".to_string(),
+            pref_value: json!("pref-value-1"),
+            feature_id: "feature-id-1".to_string(),
+            variable: "variable-1".to_string(),
+        });
+
+    // Connects to experiment slug 1
+    let pref_state_2 = GeckoPrefState::new("test.some.pref.2", Some(PrefBranch::User))
+        .with_gecko_value(json!("gecko-pref-value-2"))
+        .with_enrollment_value(PrefEnrollmentData {
+            experiment_slug: "experiment-slug-1".to_string(),
+            pref_value: json!("pref-value-2"),
+            feature_id: "feature-id-2".to_string(),
+            variable: "variable-2".to_string(),
+        });
+
+    // Other random independent experiment
+    let pref_state_3 = GeckoPrefState::new("test.some.pref.3", Some(PrefBranch::Default))
+        .with_gecko_value(json!("gecko-pref-value-3"))
+        .with_enrollment_value(PrefEnrollmentData {
+            experiment_slug: "experiment-slug-3".to_string(),
+            pref_value: json!("experiment-pref-value-3"),
+            feature_id: "feature-id-3".to_string(),
+            variable: "variable-3".to_string(),
+        });
+
+    // Experiment missing gecko value
+    let pref_state_4 = GeckoPrefState::new("test.some.pref.4", Some(PrefBranch::Default))
+        .with_enrollment_value(PrefEnrollmentData {
+            experiment_slug: "experiment-slug-4".to_string(),
+            pref_value: json!("experiment-pref-value-4"),
+            feature_id: "feature-id-4".to_string(),
+            variable: "variable-4".to_string(),
+        });
+
+    // Experiment missing enrollment data
+    let pref_state_5 = GeckoPrefState::new("test.some.pref.5", Some(PrefBranch::Default))
+        .with_gecko_value(json!("gecko-pref-value-5"));
+
+    let pref_states = vec![
+        pref_state_1,
+        pref_state_2,
+        pref_state_3,
+        pref_state_4,
+        pref_state_5,
+    ];
+    let prev_gecko_pref_states =
+        crate::stateful::gecko_prefs::build_prev_gecko_pref_states(&pref_states);
+
+    assert_eq!(3, prev_gecko_pref_states.len());
+    assert!(prev_gecko_pref_states.contains_key("experiment-slug-1"));
+    assert!(prev_gecko_pref_states.contains_key("experiment-slug-3"));
+    assert!(prev_gecko_pref_states.contains_key("experiment-slug-4"));
+
+    let PreviousGeckoPrefState {
+        original_value: original_value_1,
+        feature_id: feature_id_1,
+        variable: variable_1,
+    } = &prev_gecko_pref_states
+        .get("experiment-slug-1")
+        .expect("Missing slug")[0];
+
+    let experiment_1_states = &prev_gecko_pref_states
+        .get("experiment-slug-1")
+        .expect("Missing slug");
+    assert_eq!(2, experiment_1_states.len());
+
+    assert_eq!("feature-id-1", feature_id_1);
+    assert_eq!("variable-1", variable_1);
+
+    assert_eq!("test.some.pref.1", original_value_1.pref);
+    assert_eq!(PrefBranch::Default, original_value_1.branch);
+    assert_eq!(
+        "gecko-pref-value-1",
+        original_value_1.value.clone().unwrap()
+    );
+
+    let PreviousGeckoPrefState {
+        original_value: original_value_2,
+        feature_id: feature_id_2,
+        variable: variable_2,
+    } = &prev_gecko_pref_states
+        .get("experiment-slug-1")
+        .expect("Missing slug")[1];
+
+    assert_eq!("test.some.pref.2", original_value_2.pref);
+    assert_eq!(PrefBranch::User, original_value_2.branch);
+    assert_eq!(
+        "gecko-pref-value-2",
+        original_value_2.value.clone().unwrap()
+    );
+    assert_eq!("feature-id-2", feature_id_2);
+    assert_eq!("variable-2", variable_2);
+
+    let PreviousGeckoPrefState {
+        original_value: original_value_3,
+        feature_id: feature_id_3,
+        variable: variable_3,
+    } = &prev_gecko_pref_states
+        .get("experiment-slug-3")
+        .expect("Missing slug")[0];
+
+    let experiment_3_states = &prev_gecko_pref_states
+        .get("experiment-slug-3")
+        .expect("Missing slug");
+    assert_eq!(1, experiment_3_states.len());
+
+    assert_eq!("feature-id-3", feature_id_3);
+    assert_eq!("variable-3", variable_3);
+    assert_eq!("test.some.pref.3", original_value_3.pref);
+    assert_eq!(PrefBranch::Default, original_value_3.branch);
+    assert_eq!(
+        "gecko-pref-value-3",
+        original_value_3.value.clone().unwrap()
+    );
+
+    let PreviousGeckoPrefState {
+        original_value: original_value_4,
+        feature_id: feature_id_4,
+        variable: variable_4,
+    } = &prev_gecko_pref_states
+        .get("experiment-slug-4")
+        .expect("Missing slug")[0];
+
+    let experiment_4_states = &prev_gecko_pref_states
+        .get("experiment-slug-4")
+        .expect("Missing slug");
+    assert_eq!(1, experiment_4_states.len());
+
+    assert_eq!("feature-id-4", feature_id_4);
+    assert_eq!("variable-4", variable_4);
+    assert_eq!("test.some.pref.4", original_value_4.pref);
+    assert_eq!(PrefBranch::Default, original_value_4.branch);
+    assert_eq!(None, original_value_4.value.clone());
 
     Ok(())
 }