+++ components/nimbus/src/tests/test_enrollment_bw_compat.rs
@@ -69,3 +69,58 @@ fn test_not_enrolled_reason_schema_with_feature_conflict() {
         matches!(non_enrollment.status, EnrollmentStatus::NotEnrolled{ ref reason, ..} if reason == &NotEnrolledReason::FeatureConflict)
     );
 }
+
+// In bug 1997373, we added a `prev_gecko_pref_states` field to the EnrollmentStatus schema.
+// This test check tht the data deserializes correctly both with and without the new field.
+#[cfg(feature = "stateful")]
+#[test]
+fn test_experiment_schema_with_previous_states() {
+    // ⚠️ Warning : Do not change the JSON data used by this test. ⚠️
+    let prev_gecko_pref_states_empty: EnrollmentStatus = serde_json::from_value(json!({
+            "Enrolled": {
+            "reason": "Qualified",
+            "branch": "some_branch",
+        }
+    }))
+    .unwrap();
+    assert!(
+        matches!(prev_gecko_pref_states_empty, EnrollmentStatus::Enrolled {ref prev_gecko_pref_states, ..} if prev_gecko_pref_states.is_none())
+    );
+
+    let prev_gecko_pref_state_exists: EnrollmentStatus = serde_json::from_value(json!({
+    "Enrolled": {
+        "reason": "Qualified",
+        "branch": "some_branch",
+        "prev_gecko_pref_states": [
+        {
+            "original_value": {
+            "pref": "some_pref",
+            "branch": "default",
+            "value": 5
+            },
+            "feature_id": "some_control",
+            "variable": "some_variable"
+        },
+        {
+            "original_value": {
+            "pref": "some_pref_2",
+            "branch": "user",
+            "value": "hello"
+            },
+            "feature_id": "some_control_2",
+            "variable": "some_variable"
+        },
+        ]
+    }
+    }))
+    .unwrap();
+    assert!(matches!(
+            prev_gecko_pref_state_exists,
+            EnrollmentStatus::Enrolled {
+                prev_gecko_pref_states: Some(ref states),
+                ..
+            } if states[0].original_value.pref == "some_pref" &&  states[0].original_value.value.clone().unwrap() == 5
+            && states[0].feature_id == "some_control" &&  states[0].variable == "some_variable"
+            && states[1].original_value.pref == "some_pref_2" &&  states[1].original_value.value.clone().unwrap() == "hello"
+    ));
+}